pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
format = [
    "black>=23.3.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "jhenifer", email = "jhenifermachado211@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
